ReadMe
This repository has both the C# code as well as an executable file for the paradigm "VR hand selection". The code is commented. It should be opened within Unity3D for immediate use (version 5.5.2f was used). Unity3D can be obtained for free on their website. The executable file can be found as "_BaselineExp.zip". The source code can be found in "HandSelectionwLeap.zip".